First commit of working PIMMS using int16s for main grids to slash memory usage

+++ pimms/energy.py
@@ -338,7 +338,7 @@ class Hamiltonian:
         
         # for the 3D case
         if len(dimensions) == 3:
-            penalty = hyperloop.evaluate_angle_energy_3D(np.array(chain_positions), np.array(intcode_sequence), self.angle_lookup, num_positions)
+            penalty = hyperloop.evaluate_angle_energy_3D(np.array(chain_positions, dtype=np.int16), np.array(intcode_sequence, dtype=np.int16), self.angle_lookup, num_positions)
 
         # for the 2D case
         else:
@@ -398,9 +398,9 @@ class Hamiltonian:
 
         # initialize the residue interaction table [RIT] as a matrix
         # of zeros
-        RIT    = np.zeros(shape=(n_residues, n_residues), dtype=np.int64)
-        LRRIT  = np.zeros(shape=(n_residues, n_residues), dtype=np.int64)
-        SLRRIT = np.zeros(shape=(n_residues, n_residues), dtype=np.int64)
+        RIT    = np.zeros(shape=(n_residues, n_residues), dtype=np.int16)
+        LRRIT  = np.zeros(shape=(n_residues, n_residues), dtype=np.int16)
+        SLRRIT = np.zeros(shape=(n_residues, n_residues), dtype=np.int16)
 
         MAPPING = {}
         LR_MAPPING = {}
@@ -534,7 +534,7 @@ class Hamiltonian:
             ## while the actual int associated with the intidx reflects the identity of residue i
 
 
-            self.angle_lookup = np.zeros((int_list[-1]+1, 3, 3, 3, 3, 3, 3), dtype=np.int64)
+            self.angle_lookup = np.zeros((int_list[-1]+1, 3, 3, 3, 3, 3, 3), dtype=np.int16)
 
             AP1_count = 0
             AP2_count = 0
@@ -583,7 +583,7 @@ class Hamiltonian:
                                         self.angle_lookup[intidx, x1+1,y1+1,z1+1,x2+1,y2+1,z2+1] = penalty
                                                                             
         else:
-            self.angle_lookup = np.zeros((int_list[-1]+1, 3,3,3,3), dtype=np.int64)
+            self.angle_lookup = np.zeros((int_list[-1]+1, 3,3,3,3), dtype=np.int16)
             for x1 in range(-1,2):
                 for y1 in range(-1,2):                    
                     for x2 in range(-1,2):
